requirements.txt
flask
requests
python-docx
gunicorn

